+++ src/CountMinSketch.hpp
@@ -25,10 +25,10 @@ struct CountMinSketch {
 
 	const int r, c, m = 104729; // Yes, a magic number, I just pick a random prime
 	const int lenData;
-	unsigned long* const param1;
-	unsigned long* const param2;
-	double* const data;
-	constexpr static double infinity = std::numeric_limits<double>::infinity();
+	int* const param1;
+	int* const param2;
+	float* const data;
+	constexpr static float infinity = std::numeric_limits<float>::infinity();
 
 	// Methods
 	// --------------------------------------------------------------------------------
@@ -40,9 +40,9 @@ struct CountMinSketch {
 		r(numRow),
 		c(numColumn),
 		lenData(r * c),
-		param1(new unsigned long[r]),
-		param2(new unsigned long[r]),
-		data(new double[lenData]) {
+		param1(new int[r]),
+		param2(new int[r]),
+		data(new float[lenData]) {
 		for (int i = 0; i < r; i++) {
 			param1[i] = rand() + 1; // ×0 is not a good idea, see Hash()
 			param2[i] = rand();
@@ -54,9 +54,9 @@ struct CountMinSketch {
 		r(b.r),
 		c(b.c),
 		lenData(b.lenData),
-		param1(new unsigned long[r]),
-		param2(new unsigned long[r]),
-		data(new double[lenData]) {
+		param1(new int[r]),
+		param2(new int[r]),
+		data(new float[lenData]) {
 		std::copy(b.param1, b.param1 + r, param1);
 		std::copy(b.param2, b.param2 + r, param2);
 		std::copy(b.data, b.data + lenData, data);
@@ -68,36 +68,36 @@ struct CountMinSketch {
 		delete[] data;
 	}
 
-	void ClearAll(double with = 0) const {
+	void ClearAll(float with = 0) const {
 		std::fill(data, data + lenData, with);
 	}
 
-	void MultiplyAll(double by) const {
+	void MultiplyAll(float by) const {
 		for (int i = 0, I = lenData; i < I; i++) // Vectorization
 			data[i] *= by;
 	}
 
-	void Hash(unsigned long* indexOut, unsigned long a, unsigned long b = 0) const {
+	void Hash(int* indexOut, int a, int b = 0) const {
 		for (int i = 0; i < r; i++) {
 			indexOut[i] = ((a + m * b) * param1[i] + param2[i]) % c;
 			indexOut[i] += i * c + (indexOut[i] < 0 ? c : 0);
 		}
 	}
 
-	double operator()(const unsigned long* index) const {
-		double least = infinity;
+	float operator()(const int* index) const {
+		float least = infinity;
 		for (int i = 0; i < r; i++)
 			least = std::min(least, data[index[i]]);
 		return least;
 	}
 
-	double Assign(const unsigned long* index, double with) const {
+	float Assign(const int* index, float with) const {
 		for (int i = 0; i < r; i++)
 			data[index[i]] = with;
 		return with;
 	}
 
-	void Add(const unsigned long* index, double by = 1) const {
+	void Add(const int* index, float by = 1) const {
 		for (int i = 0; i < r; i++)
 			data[index[i]] += by;
 	}
